Backend/requirements.txt
sentence-transformers==2.2.2
langchain>=0.2.15
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone>=0.1.2
langchain_community
langchain_openai
langchain_experimental
Groq
joblib
flask_cors
mistralai
tensorflow
whisper
SpeechRecognition
gunicorn
